+++ src/main/editor/Prefabs.java
@@ -17,7 +17,7 @@ import java.util.List;
 import static main.Configuration.*;
 
 public class Prefabs {
-    private static GameObject generateObject(float width, float height){
+    private static GameObject generateObject(double width, double height){
         GameObject holdingObject = new GameObject("Generated");
         holdingObject.addComponent(new Transform(new Vector2d(), new Vector2d(width * spriteSize, height * spriteSize)));
         holdingObject.setTransformFromItself();
@@ -25,7 +25,7 @@ public class Prefabs {
         return holdingObject;
     }
 
-    public static GameObject generateSpriteObject(SpriteRenderer spriteR, float width, float height){
+    public static GameObject generateSpriteObject(SpriteRenderer spriteR, double width, double height){
         GameObject holdingObject = generateObject(width, height);
         SpriteRenderer spriteRenderer = new SpriteRenderer(spriteR.getTexture(), width, height, spriteR.getSpriteCords());
         holdingObject.addComponent(spriteRenderer);
@@ -33,7 +33,7 @@ public class Prefabs {
         return holdingObject;
     }
 
-    public static GameObject generateSolidObject(SpriteRenderer spriteR, float width, float height){
+    public static GameObject generateSolidObject(SpriteRenderer spriteR, double width, double height){
         GameObject solidObject = generateSpriteObject(spriteR, width, height);
         RigidBody rigidBody = new RigidBody();
         rigidBody.setBodyType(BodyType.Static);
@@ -44,7 +44,7 @@ public class Prefabs {
         return solidObject;
     }
 
-    public static GameObject generateAnimateObject(float width, float height, float defaultFrameTime, String title, List<SpriteRenderer> spriteRenderList){
+    public static GameObject generateAnimateObject(double width, double height, double defaultFrameTime, String title, List<SpriteRenderer> spriteRenderList){
         Animation animation = new Animation(title);
         for(SpriteRenderer spriteRenderer: spriteRenderList){
             animation.addFrame(spriteRenderer, defaultFrameTime);
@@ -53,7 +53,7 @@ public class Prefabs {
         return  generateAnimateObject(width, height, animation);
     }
 
-    public static GameObject generateAnimateObject(float width, float height, Animation animation){
+    public static GameObject generateAnimateObject(double width, double height, Animation animation){
         GameObject animatedHoldingObject = generateObject(width, height);
         StateMachine stateMachine = new StateMachine();
         stateMachine.addState(animation);
